+++ CHANGES.md
@@ -1,5 +1,5 @@
 
-
+- Uchar.t support. At the API level only `Jsonm.error` changes.
 - Build depend on topkg.
 - Relicensed from BSD3 to ISC.
 - Fix `Jsonm.decode` not eventually returning `End on toplevel